Results functionality, ironing out bugs relating to deletion, fixed navigationlinks

+++ src/main/java/wad/service/EmitService.java
@@ -1,4 +1,3 @@
-
 package wad.service;
 
 import java.text.SimpleDateFormat;
@@ -14,99 +13,103 @@ import wad.repository.EmitRepository;
 
 @Service
 public class EmitService {
-    
+
     @Autowired
     private EmitRepository emitRepository;
-    
+
     @Autowired
     private EmitPunchService emitPunchService;
-    
+
     @Autowired
     private CompetitorService competitorService;
-    
-    public void saveEmit(Emit emit){
+
+    public void saveEmit(Emit emit) {
         emitRepository.save(emit);
     }
-    
-    public List<Emit> getEmits(){
+
+    public List<Emit> getEmits() {
         return emitRepository.findAll();
     }
-    
-    public Emit getEmit(Long id){
+
+    public Emit getEmit(Long id) {
         return emitRepository.findOne(id);
     }
-    
-    public void deleteEmit(Long id){
+
+    public void deleteEmit(Long id) {
         Emit emit = emitRepository.findOne(id);
-        
+
         Competitor owner = emit.getOwner();
-        owner.setEmit(null);
-        owner.setEmitNumber(null);
-        competitorService.saveCompetitor(owner);
-        
+        if (owner != null) {
+            owner.setEmit(null);
+            owner.setEmitNumber(null);
+            competitorService.saveCompetitor(owner);
+        }
+
         List<EmitPunch> punches = emit.getEmitPunches();
-        emit.setEmitPunches(null);
-        for(EmitPunch punch : punches){
-            emitPunchService.deleteEmitPunch(punch.getId());
+        if (punches != null) {
+            emit.setEmitPunches(null);
+            for (EmitPunch punch : punches) {
+                emitPunchService.deleteEmitPunch(punch.getId());
+            }
         }
         emitRepository.delete(id);
-    }   
-    
-    public Emit getByNumber(String number){
+    }
+
+    public Emit getByNumber(String number) {
         return emitRepository.findByNumber(number);
     }
-    
-    public List<String> processEmitPunchesToCodes(Emit emit){
+
+    public List<String> processEmitPunchesToCodes(Emit emit) {
         List<EmitPunch> punches = emit.getEmitPunches();
         List<String> codes = new ArrayList<>();
-        for(EmitPunch punch : punches){
+        for (EmitPunch punch : punches) {
             codes.add(punch.getPunchCode());
         }
         return codes;
     }
-    
-    public List<String> processEmitPunchesToSplits(Emit emit){
+
+    public List<String> processEmitPunchesToSplits(Emit emit) {
         List<EmitPunch> punches = emit.getEmitPunches();
         List<String> splits = new ArrayList<>();
-        
-        for(int i = 0; i < punches.size()-1;i++){
-            Long split = punches.get(i+1).getTimestamp().getTime()-punches.get(i).getTimestamp().getTime();
+
+        for (int i = 0; i < punches.size() - 1; i++) {
+            Long split = punches.get(i + 1).getTimestamp().getTime() - punches.get(i).getTimestamp().getTime();
             splits.add(millisecondsToMinutesSeconds(split));
         }
-        
+
         return splits;
-        
+
     }
-    
-    public List<String> processEmitPunchesToCumulativeTime(Emit emit){
+
+    public List<String> processEmitPunchesToCumulativeTime(Emit emit) {
         List<EmitPunch> punches = emit.getEmitPunches();
         List<String> cumulative = new ArrayList<>();
         Long cumulativeSplit = 0L;
-        for(int i = 0; i < punches.size()-1;i++){
-            cumulativeSplit += punches.get(i+1).getTimestamp().getTime()-punches.get(i).getTimestamp().getTime();
+        for (int i = 0; i < punches.size() - 1; i++) {
+            cumulativeSplit += punches.get(i + 1).getTimestamp().getTime() - punches.get(i).getTimestamp().getTime();
             cumulative.add(millisecondsToMinutesSeconds(cumulativeSplit));
         }
-        
+
         return cumulative;
     }
-    
-    public List<String> processEmitPunchesToPrintableSplits(Emit emit){
+
+    public List<String> processEmitPunchesToPrintableSplits(Emit emit) {
         List<EmitPunch> punches = emit.getEmitPunches();
         List<String> printableSplits = new ArrayList<>();
         EmitPunch currentPunch;
         Long cumulativeSplit = 0L;
-        for(int i = 0; i < punches.size()-1;i++){
-            currentPunch = punches.get(i+1);
-            cumulativeSplit += punches.get(i+1).getTimestamp().getTime()-punches.get(i).getTimestamp().getTime();
-            Long split = punches.get(i+1).getTimestamp().getTime()-punches.get(i).getTimestamp().getTime();
-            printableSplits.add(millisecondsToMinutesSeconds(split)+ " " + millisecondsToMinutesSeconds(cumulativeSplit) + " " + currentPunch.getPunchCode());
+        for (int i = 0; i < punches.size() - 1; i++) {
+            currentPunch = punches.get(i + 1);
+            cumulativeSplit += punches.get(i + 1).getTimestamp().getTime() - punches.get(i).getTimestamp().getTime();
+            Long split = punches.get(i + 1).getTimestamp().getTime() - punches.get(i).getTimestamp().getTime();
+            printableSplits.add(millisecondsToMinutesSeconds(split) + " " + millisecondsToMinutesSeconds(cumulativeSplit) + " " + currentPunch.getPunchCode());
         }
-        
+
         return printableSplits;
-        
+
     }
-    
-    public String millisecondsToMinutesSeconds(Long ms){
+
+    public String millisecondsToMinutesSeconds(Long ms) {
         return new SimpleDateFormat("mm:ss").format(new Date(ms));
     }
 }
